pyproject.toml
[tool.poetry]
name = "jeopardy-api"
version = "0.1.0"
description = "Jeopardy API"
authors = ["Nurlat Bekdullayev <natuspati@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
asyncpg = "^0.29.0"
uvicorn = "^0.30.6"
pydantic-settings = "^2.4.0"
pytz = "^2024.1"
orjson = "^3.10.7"
alembic = "^1.13.2"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
wemake-python-styleguide = "^0.19.2"
autoflake = "^2.3.1"
click = "^8.1.7"
flake8-pytest-style = "^2.0.0"
pytest-asyncio = "0.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest_env]
JEOPARDY_ENVIRONMENT = "test"
JEOPARDY_DB_HOST = "localhost"
JEOPARDY_DB_PORT = "5432"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["src",]
